style: formatação de código

+++ main.py
@@ -17,7 +17,7 @@ CATEGORIAS = [
 ]
 
 MESES = {
-    "Janeiro": "01", 
+    "Janeiro": "01",
     "Fevereiro": "02",
     "Março": "03",
     "Abril": "04",
@@ -31,19 +31,21 @@ MESES = {
     "Dezembro": "12",
 }
 
+
 # Função para formatar a data
 def formatar_data(data):
     dia, mes, ano = data.split("/")
     return f"{dia}/{MESES[mes]}/{ano}"
 
+
 def adicionar_despesa(descricao, valor, data, categoria):
-    with open("gastos.csv", "a", newline="", encoding='utf-8') as file:
+    with open("gastos.csv", "a", newline="", encoding="utf-8") as file:
         writer = csv.writer(file)
         writer.writerow([descricao, valor, data, categoria])
 
 
 def editar_despesa(index, descricao, valor, data, categoria):
-    with open("gastos.csv", "r", newline="", encoding='utf-8') as file:
+    with open("gastos.csv", "r", newline="", encoding="utf-8") as file:
         rows = list(csv.reader(file))
     rows[index] = [descricao, valor, data, categoria]
     with open("gastos.csv", "w", newline="") as file:
@@ -52,7 +54,7 @@ def editar_despesa(index, descricao, valor, data, categoria):
 
 
 def excluir_despesa(index):
-    with open("gastos.csv", "r", newline="", encoding='utf-8') as file:
+    with open("gastos.csv", "r", newline="", encoding="utf-8") as file:
         rows = list(csv.reader(file))
     del rows[index]
     with open("gastos.csv", "w", newline="") as file:
@@ -61,68 +63,78 @@ def excluir_despesa(index):
 
 
 def filtrar_despesas(mes, categoria):
-    with open('gastos.csv', 'rb') as file:
+    with open("gastos.csv", "rb") as file:
         result = chardet.detect(file.read())
-    encoding = result['encoding']
+    encoding = result["encoding"]
 
-    with open('gastos.csv', 'r', newline='', encoding=encoding) as file:
+    with open("gastos.csv", "r", newline="", encoding=encoding) as file:
         reader = csv.reader(file)
         rows = list(reader)
 
-    df = pd.DataFrame(rows, columns=['Descrição', 'Valor', 'Data', 'Categoria'])
-    df['Data'] = pd.to_datetime(df['Data'])
+    df = pd.DataFrame(rows, columns=["Descrição", "Valor", "Data", "Categoria"])
+    df["Data"] = pd.to_datetime(df["Data"])
 
     if mes:
-        df = df[df['Data'].dt.month == int(mes)]
+        df = df[df["Data"].dt.month == int(mes)]
     if categoria:
-        df = df[df['Categoria'] == categoria]
+        df = df[df["Categoria"] == categoria]
 
     if not df.empty:
-        df['Data'] = df['Data'].dt.strftime('%d/%m/%Y')
-        df = df.rename(columns={'Descrição': 'Descrição', 'Valor': 'Valor', 'Data': 'Data', 'Categoria': 'Categoria'})
-        return df.to_dict('records')
+        df["Data"] = df["Data"].dt.strftime("%d/%m/%Y")
+        df = df.rename(
+            columns={
+                "Descrição": "Descrição",
+                "Valor": "Valor",
+                "Data": "Data",
+                "Categoria": "Categoria",
+            }
+        )
+        return df.to_dict("records")
 
     return []
 
 
 def gerar_estatisticas():
-    with open('gastos.csv', 'rb') as file:
+    with open("gastos.csv", "rb") as file:
         result = chardet.detect(file.read())
-    encoding = result['encoding']
+    encoding = result["encoding"]
 
-    with open('gastos.csv', 'r', newline='', encoding=encoding) as file:
+    with open("gastos.csv", "r", newline="", encoding=encoding) as file:
         reader = csv.reader(file)
         rows = list(reader)
-    df = pd.DataFrame(rows, columns=['Descrição', 'Valor', 'Data', 'Categoria'])
-    df['Valor'] = df['Valor'].astype(float)
+    df = pd.DataFrame(rows, columns=["Descrição", "Valor", "Data", "Categoria"])
+    df["Valor"] = df["Valor"].astype(float)
     valores = [float(row[1]) for row in rows]
     total = sum(valores)
     media = total / len(valores)
     maior_valor = max(valores)
 
-    st.subheader('Total de Gastos')
-    total = df['Valor'].sum()
+    st.subheader("Total de Gastos")
+    total = df["Valor"].sum()
     menor_valor = min(valores)
-    st.write(f'Total: R$ {total:.2f}')
-    st.write(f'Média: R$ {media:.2f}')
-    st.write(f'Maior valor: R$ {maior_valor:.2f}')
-    st.write(f'Menor valor: R$ {menor_valor:.2f}')
+    st.write(f"Total: R$ {total:.2f}")
+    st.write(f"Média: R$ {media:.2f}")
+    st.write(f"Maior valor: R$ {maior_valor:.2f}")
+    st.write(f"Menor valor: R$ {menor_valor:.2f}")
 
-    st.subheader('Gastos por Categoria')
-    gastos_por_categoria = df.groupby('Categoria')['Valor'].sum()
+    st.subheader("Gastos por Categoria")
+    gastos_por_categoria = df.groupby("Categoria")["Valor"].sum()
     st.bar_chart(gastos_por_categoria)
 
-    st.subheader('Média de Gastos por Categoria')
-    media_por_categoria = df.groupby('Categoria')['Valor'].mean()
+    st.subheader("Média de Gastos por Categoria")
+    media_por_categoria = df.groupby("Categoria")["Valor"].mean()
     st.bar_chart(media_por_categoria)
 
-    st.subheader('Visualização dos Gastos por Categoria')
+    st.subheader("Visualização dos Gastos por Categoria")
     fig, ax = plt.subplots()
-    ax.pie(gastos_por_categoria, labels=gastos_por_categoria.index, autopct='%1.1f%%', startangle=90)
-    ax.axis('equal')
+    ax.pie(
+        gastos_por_categoria,
+        labels=gastos_por_categoria.index,
+        autopct="%1.1f%%",
+        startangle=90,
+    )
+    ax.axis("equal")
     st.pyplot(fig)
-
-
 
 
 def main():
@@ -163,7 +175,9 @@ def main():
                 format="%.2f",
                 value=float(despesas[selected_index][1]),
             )
-            data_visual = datetime.strptime(despesas[selected_index][2], '%Y-%m-%d').date()
+            data_visual = datetime.strptime(
+                despesas[selected_index][2], "%Y-%m-%d"
+            ).date()
             data = st.date_input("Data", value=data_visual)
             categoria = st.selectbox(
                 "Categoria",
